fix: respect specified config filename
Previously, the load_config_at_path function removed the base filename
and searched for filenames listed in filename_options, such as .sqruff
and .sqlfluff. This fix changes the logic so that the filename_options
list is only used if the file path is an existing directory. If instead
the file path is an existing file, then simply load the config from the
file.

fixes #1131