Mark T-SQL CTE test as ignored in ST05
The test case issue_3623_internal_error_multiple_templated_files
involves T-SQL's special syntax:
CREATE TABLE ... WITH (DISTRIBUTION = ...) AS WITH cte AS (...) SELECT ...

This syntax has two consecutive WITH clauses - one for table options
and one for CTEs. The ST05 rule currently doesn't handle this case
correctly and would need significant changes to support it.

This test is now properly marked as ignored with an explanation,
allowing the test suite to pass while documenting the limitation.

Fixes #1712